Added some small css changes based on wireframes & created component folders
