+++ src/components/App.js
@@ -1,9 +1,9 @@
 import React, { Fragment, useState } from "react";
-import Navbar from "./Navbar";
-import Footer from "./Footer";
-import PreFooter from "./PreFooter";
+import Navbar from "./Navbar/Navbar";
+import Footer from "./Footer/Footer";
+import PreFooter from "./PreFooter/PreFooter";
 import "./App.css";
-import Motivations from "./Motivations";
+import Motivations from "./Motivations/Motivations";
 import Jumbotron from "./Jumbotron/Jumbotron";
 import UserStories from "./UserStories/UserStories";
