+++ src/components/Footer/Footer.css
@@ -3,6 +3,7 @@
   justify-content: center;
   height: 194px;
   background-color: #0f2437;
+  letter-spacing: 0.01em;
 }
 
 .footer__logo {
