+++ src/components/Navbar/Navbar.css
@@ -1,6 +1,5 @@
 .container {
   display: flex;
-  /*width: 1440px;*/
   width: 99vw;
 }
 
@@ -8,6 +7,7 @@
   display: flex;
   justify-content: center;
   padding: 14px;
+  letter-spacing: 0.01em;
 }
 
 .navbar__logo {
