+++ src/components/UserStories/UserStories.css
@@ -7,6 +7,7 @@
   align-items: center;
   padding: 80px 0;
   background: white;
+  letter-spacing: 0.01em;
 }
 
 .user-stories {
@@ -38,7 +39,7 @@
   justify-content: space-between;
   align-items: center;
   margin: 5px 0;
-  padding: 1rem 2rem;
+  padding: 1rem 2.5rem;
 }
 
 .stories-nav-btn {
@@ -62,14 +63,13 @@
   line-height: 162.2%;
 }
 .stories-header {
-  font-family: "Open Sans";
-  font-style: normal;
   font-weight: bold;
-  font-size: 26px;
+  font-size: 32px;
   line-height: 44px;
   letter-spacing: 0.01em;
-  padding: 20px 0;
-  margin: 0 0 50px 0;
+  margin-bottom: 108px;
+  margin-top: 0;
+  color: #333333;
 }
 .nav-circles-container {
   display: flex;
